feat(state)!: move TanStack DB surface to @durable-streams/state/db subpath
#381 made `@tanstack/db` a peer dependency of `@durable-streams/state`
(correctly — it's a singleton, so two copies in a tree don't interoperate).
But the package's main entry eagerly re-exported ~30 runtime values from
`@tanstack/db`, so importing *anything* from `@durable-streams/state` forced
that peer to be resolvable at module-load time.

That broke publishable dependents that don't use the reactive layer. The
clearest case is `@durable-streams/server`, which imports only
`createStateSchema` (pure schema validation, no `@tanstack/db` at runtime) yet
crashed on import with `ERR_MODULE_NOT_FOUND: @tanstack/db` for any consumer
who hadn't installed the unmet peer (npm auto-installs peers, but pnpm/yarn do
not).

This splits the package along the protocol's natural write/read boundary:

- `@durable-streams/state` (main entry) is now free of `@tanstack/db`. It
  exposes the producer/backend surface: `createStateSchema` + change-event
  helpers, `MaterializedState`, and the state-event types/guards. Backends like
  the server can depend on it without the peer.
- `@durable-streams/state/db` (new subpath) holds the reactive, TanStack
  DB-backed layer: `createStreamDB`, `getStreamDBCollectionId`, the
  `StreamDB*`/`Action*` types, and the `@tanstack/db` convenience re-exports.
  It is a strict superset of the main entry, so reactive consumers only change
  the import path.

Mechanics:
- Extract the db-free schema/event-construction code into `src/schema.ts`.
- `src/index.ts` re-exports only the db-free surface; `src/db.ts` is the new
  superset entry; `stream-db.ts` re-exports `createStateSchema` for back-compat.
- Add the `./db` export map entry and a second tsdown build entry.
- Update the four examples that use `createStreamDB` to import from `/db`
  (one-line path change each).

Tests:
- New `test/db-free-entry.test.ts` traces each entry's module graph and asserts
  the main entry never imports `@tanstack/db` while `/db` does. This failed
  before the split and passes now.
- Verified the built `server` imports cleanly with `@tanstack/db` removed from
  the tree (previously `ERR_MODULE_NOT_FOUND`).

BREAKING CHANGE: `createStreamDB`, `getStreamDBCollectionId`, the
StreamDB/Action types, and the `@tanstack/db` re-exports are no longer exported
from `@durable-streams/state` — import them from `@durable-streams/state/db`.
The db-free APIs are unchanged on the main entry.
